renamed folders to kebab-case
