+++ apps/admin-x-settings/src/components/settings/growth/embed-signup/embed-signup-sidebar.tsx
