+++ apps/admin-x-settings/src/components/settings/membership/member-emails.tsx
@@ -2,7 +2,7 @@ import FakeLogo from '../../../assets/images/explore-default-logo.png';
 import NiceModal from '@ebay/nice-modal-react';
 import React, {useState} from 'react';
 import TopLevelGroup from '../../top-level-group';
-import WelcomeEmailModal from './memberEmails/welcome-email-modal';
+import WelcomeEmailModal from './member-emails/welcome-email-modal';
 import {Button, Separator, SettingGroupContent, Toggle, withErrorBoundary} from '@tryghost/admin-x-design-system';
 import {getSettingValues} from '@tryghost/admin-x-framework/api/settings';
 import {useGlobalData} from '../../providers/global-data-provider';
